requirements.txt
Django>=3.0,<4.0
psycopg2
djangorestframework
django-filter